+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+## [0.7.0] - 2022-06-01
 ### Added
 * `Plotter.raster()` and `raster_mut()` (returning reference)
 ### Changed